Refactor strings into arguments when deleting/inserting into a table (#41)
diskquota used long strings when deleting/inserting into the
diskquota.table_size table. This resulted in high memory consumption, both for
constructing such long strings and for parsing them. This patch reworks the
logic to use arguments consisting of arrays. As a result, significantly less
memory is spent, since the query itself is very short and there is no need to
waste memory on constructing and parsing it, array arguments are passed as is.

This patch also fixes the error duplicate key value violates unique constraint
"table_size_pkey". In the flush_to_table_size function, it may happen that when
updating the table size, the delete goes to one batch and the insert goes to
another, and the insert is performed earlier and a duplicate error occurs.

Ticket: ADBDEV-6520